Reformatting, moving code (#149)
